Fix bug when at least one message has not failed
GET on a non existent key e.g. `stats:failed` (when no messages have
failed yet) returns a redis.Nil error, preventing the other metrics
to be collected.

This is mostly not an issue with an in-use production system, but
means that metrics are not collected properly from a new environment
until there has been a failure.